Update the res.partner model and view to customize it for the portal.
The res.partner view is set to know if the user is from the portal
and then to hide some extra fields.
On top of that is_company is set to True at the creation of a new record
if the user that created the record is a portal user.

To finish, a bit of polish on the color of required field color, to
make the view a bit lighter to watch.