M23-iframe take2 (#724)
* res_partner_count: add host and ip to the visit

* website_iframe_host: simplification of the condition of authorization

simplification. No white list. Only data in the database. if the host is
not in the database, it is not allowed.

* website_iframe_host: add localhost and cgstudiomap as hosts

* website_iframe_host: fix issue with navbar and add light_hosting

* with the previous logic, the navbar was hidden by default, while we
wanted the opposite. It would have oblige us to manage all the pages.
Now the navbar is on by default, so the pages should not be affected.
* light_hosting tags a host in a way it won't host pages require login
(add studio, edit studio mainly targeted here)